add partial code of tpch dbgen
