+++ tpch/dbgen/driver_test.go
@@ -1,0 +1,15 @@
+package dbgen
+
+import (
+	"os"
+	"testing"
+)
+
+func TestMain(m *testing.M) {
+	initDriver(1)
+	os.Exit(m.Run())
+}
+
+//func TestGenTable(t *testing.T) {
+//	genTable(ORDER, 1, 10)
+//}
